requirements.txt
boto3==1.1.3
Django==1.8.9
djangorestframework==3.1.2
django-annoying==0.8.1
django-celery-with-redis==3.0
django-cors-headers==1.1.0
django-extensions==1.5.5
django-heroku-memcacheify==0.8
django-heroku-postgresify==0.4
django-localflavor==1.1
django-nose==1.4
django-ses-backend==0.1.1
-e git+https://github.com/jschneier/django-storages-redux.git@f2880b16b57a36b241a54932255e1a852c0e0ac7#egg=django_storages_redux
djrill==1.4.0
-e git+https://github.com/firebase/firebase-token-generator-python@f4197b1fb8108c6495817cf6e4cf5443dbd7f410#egg=firebase_token_generator
kombu==3.0.33
flake8==2.4.1
newrelic==2.56.0.42
gunicorn==19.3.0
piprot==0.9.6
Pillow==3.1.1
polytester==1.2.0
psycopg2==2.6
python-firebase==1.2
rollbar==0.9.11
sorl-thumbnail==12.2
simple-crypt==4.1.0

# Dev only
ipython==3.1.0
## The following requirements were added by pip freeze:
amqp==1.4.6
anyjson==0.3.3
argh==0.26.1
args==0.1.0
billiard==3.3.0.20
boto==2.38.0
botocore==1.2.4
celery==3.1.18
cffi==1.5
clint==0.4.1
cryptography==1.2.2
dj-database-url==0.3.0
django-celery==3.1.16
django-pylibmc==0.5.0
-e git+https://github.com/jschneier/django-storages-redux.git@7875a755f07009f0e6885280d2a782269bb8c258#egg=django_storages
docutils==0.12
futures==2.2.0
gnureadline==6.3.3
idna==2.0
jmespath==0.7.1
mccabe==0.3
nose==1.3.7
pathtools==0.1.2
pep8==1.5.7
pyasn1==0.1.7
pycparser==2.13
pycrypto==2.6.1
pyflakes==0.8.1
pylibmc==1.4.3
python-dateutil==2.4.2
pytz==2015.4
PyYAML==3.11
redis==2.10.3
requests==2.7.0
requests-futures==0.9.5
six==1.9.0
virtualenv==14.0.6
watchdog==0.8.2
